Distortion model code cleaned up, separated original fisheye application and the upscaled application. Added a section in the READMe detailing these.
